requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2
dotenv
gunicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart